+++ packages/quicktype-core/src/language/Php/index.ts
@@ -1,0 +1,2 @@
+export { PhpTargetLanguage, phpOptions } from "./language";
+export { PhpRenderer } from "./PhpRenderer";
