+++ packages/quicktype-core/src/language/Php/language.ts
@@ -1,0 +1,46 @@
+import * as _ from "lodash";
+
+import { type RenderContext } from "../../Renderer";
+import { BooleanOption, type Option, getOptionValues } from "../../RendererOptions";
+import { AcronymStyleOptions, acronymOption } from "../../support/Acronyms";
+import { TargetLanguage } from "../../TargetLanguage";
+import { type PrimitiveStringTypeKind, type TransformedStringTypeKind } from "../../Type";
+import { type StringTypeMapping } from "../../TypeBuilder";
+import { type FixMeOptionsAnyType, type FixMeOptionsType } from "../../types";
+
+import { PhpRenderer } from "./PhpRenderer";
+
+export const phpOptions = {
+    withGet: new BooleanOption("with-get", "Create Getter", true),
+    fastGet: new BooleanOption("fast-get", "getter without validation", false),
+    withSet: new BooleanOption("with-set", "Create Setter", false),
+    withClosing: new BooleanOption("with-closing", "PHP Closing Tag", false),
+    acronymStyle: acronymOption(AcronymStyleOptions.Pascal)
+};
+export class PhpTargetLanguage extends TargetLanguage {
+    public constructor() {
+        super("PHP", ["php"], "php");
+    }
+
+    protected getOptions(): Array<Option<FixMeOptionsAnyType>> {
+        return _.values(phpOptions);
+    }
+
+    public get supportsUnionsWithBothNumberTypes(): boolean {
+        return true;
+    }
+
+    protected makeRenderer(renderContext: RenderContext, untypedOptionValues: FixMeOptionsType): PhpRenderer {
+        const options = getOptionValues(phpOptions, untypedOptionValues);
+        return new PhpRenderer(this, renderContext, options);
+    }
+
+    public get stringTypeMapping(): StringTypeMapping {
+        const mapping: Map<TransformedStringTypeKind, PrimitiveStringTypeKind> = new Map();
+        mapping.set("date", "date"); // TODO is not implemented yet
+        mapping.set("time", "time"); // TODO is not implemented yet
+        mapping.set("uuid", "uuid"); // TODO is not implemented yet
+        mapping.set("date-time", "date-time");
+        return mapping;
+    }
+}
